Hoist await and yield expressions
The helper wraps the original argument with a new function. This is a
problem if the argument contains yield or await expressions, since the
function they now run is is not a generator or marked as asynchronous.

Instead hoist these expressions so they're resolved outside of the
helper.

Fixes https://github.com/avajs/ava/issues/987.